Move randStringBytes to helpers
